Revamp triggers to be more general
This splits the handling of triggers in the config file and triggers
in used in the code up into two enums, the former one being private.

This simplifies parsing the config file, while not exposing useless
stuff to modules.

The format of the config file is tweaked to be more explicit, e.g.
what as previously "/me rolls" is now "<action rolls>".

As an assorted change lazy-static is removed, and instead the regex'
are made part of the state of the modules, and the URL regex is replaced
by a nifty parser crate.

As an additional assorted change, the cargo.lock file of the core
crate is removed, as one shouldn't include these for crates.